Adapt Virtio to function seamlessly with ThreadX
The most significant changes occurred in `rpmsg_char.c`, which is the
platform driver for the rpmsg framework and the Virtio device registration.
Firstly, I modified the number of ring buffers to match the ThreadX Virtio ring.
Secondly, I rewrote the ISR to issue the new `delay_work` to call the
Virtqueue callback function.

P.S.: I will rename the file `rpmsg_char.c` in the next commit.